moved check for deletable subs inside delete event type tx

+++ api-metastore/src/main/java/org/zalando/nakadi/service/EventTypeService.java
@@ -1,6 +1,5 @@
 package org.zalando.nakadi.service;
 
-import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Multimap;
 import org.json.JSONObject;
 import org.slf4j.Logger;
@@ -345,12 +344,7 @@ public class EventTypeService {
                         "You cannot delete event-type without authorization.");
             }
 
-            if (featureToggleService.isFeatureEnabled(DELETE_EVENT_TYPE_WITH_SUBSCRIPTIONS) ||
-                    !hasNonDeletableSubscriptions(eventType.getName())) {
-                topicsToDelete = deleteEventTypeWithSubscriptions(eventTypeName);
-            } else {
-                throw new ConflictException("Can't remove event type " + eventTypeName + ", as it has subscriptions");
-            }
+            topicsToDelete = deleteEventTypeWithSubscriptions(eventTypeName);
             eventTypeCache.invalidate(eventTypeName);
         } catch (final InterruptedException e) {
             Thread.currentThread().interrupt();
@@ -407,6 +401,12 @@ public class EventTypeService {
                         subscriptionRepository.listSubscriptions(Set.of(eventType), Optional.empty(),
                                 Optional.empty(), Optional.empty());
 
+                if (!(featureToggleService.isFeatureEnabled(DELETE_EVENT_TYPE_WITH_SUBSCRIPTIONS)
+                        || onlyDeletableSubscriptions(subscriptions))) {
+                    throw new ConflictException("Can't remove event type " + eventType +
+                            ", as it has subscriptions");
+                }
+
                 subscriptions.forEach(s -> {
                         try {
                             subscriptionRepository.deleteSubscription(s.getId());
@@ -422,24 +422,11 @@ public class EventTypeService {
         }
     }
 
-
-    // TODO: This method should be fixed by creating proper db query.
-    private boolean hasNonDeletableSubscriptions(final String eventTypeName) {
-
-        SubscriptionTokenLister.ListResult list = subscriptionTokenLister.listSubscriptions(
-                ImmutableSet.of(eventTypeName), Optional.empty(), Optional.empty(), null, 20);
-        while (null != list) {
-            for (final Subscription sub : list.getItems()) {
-                if (!sub.getConsumerGroup().equals(nakadiSettings.getDeletableSubscriptionConsumerGroup())
-                        || !sub.getOwningApplication()
-                        .equals(nakadiSettings.getDeletableSubscriptionOwningApplication())) {
-                    return true;
-                }
-            }
-            list = null == list.getNext() ? null : subscriptionTokenLister.listSubscriptions(
-                    ImmutableSet.of(eventTypeName), Optional.empty(), Optional.empty(), list.getNext(), 20);
-        }
-        return false;
+    private boolean onlyDeletableSubscriptions(final List<Subscription> subscriptions) {
+        final String owningApplication = nakadiSettings.getDeletableSubscriptionOwningApplication();
+        final String consumerGroup = nakadiSettings.getDeletableSubscriptionConsumerGroup();
+        return subscriptions.stream().allMatch(sub ->
+                sub.getOwningApplication().equals(owningApplication) && sub.getConsumerGroup().equals(consumerGroup));
     }
 
     public void update(final String eventTypeName,
